core: stdcm: remove previous version
