+++ core/src/main/java/fr/sncf/osrd/api/stdcm/OccupancyBlock.java
@@ -1,7 +1,8 @@
-package fr.sncf.osrd.api.stdcm.new_pipeline;
+package fr.sncf.osrd.api.stdcm;
 
 /** The given element is unavailable from timeStart until timeEnd,
- * in the space between distanceStart and distanceEnd. */
+ * in the space between distanceStart and distanceEnd.
+ * Distances are relative to the start of the route. */
 public record OccupancyBlock(
         double timeStart,
         double timeEnd,
