+++ core/src/main/java/fr/sncf/osrd/api/stdcm/STDCMResult.java
@@ -1,4 +1,4 @@
-package fr.sncf.osrd.api.stdcm.new_pipeline;
+package fr.sncf.osrd.api.stdcm;
 
 import fr.sncf.osrd.envelope.Envelope;
 import fr.sncf.osrd.envelope_sim.PhysicsPath;
@@ -6,8 +6,9 @@ import fr.sncf.osrd.infra.api.signaling.SignalingRoute;
 import fr.sncf.osrd.infra_state.api.TrainPath;
 import fr.sncf.osrd.utils.graph.Pathfinding;
 
-/** This is the result of the STDCM computation
- * made of a physical path part and and envelope. */
+/** This is the result of the STDCM computation.
+ * It is made of a physical path part and and envelope, as well as different representations
+ * of the same data that can be reused in later steps. */
 public record STDCMResult(
         Pathfinding.Result<SignalingRoute> routes,
         Envelope envelope,
